# src/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "node-mvc",
        "version": "1.0.0",
        "title": "node-mvc",
        "contact": {
            "email": "xxxxxxxxxxxm@163.com"
        }
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "tags": [
        {
            "name": "User",
            "description": "关于用户类的操作"
        },
        {
            "name": "Common",
            "description": "通用的一类操作"
        }
    ],
    "schemes": [
        "http"
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Common"
                ],
                "summary": "注册",
                "description": "用户注册",
                "operationId": "register",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "创建一个用户对象",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/definitions/User"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "get": {
                "tags": [
                    "Common"
                ],
                "summary": "登录",
                "description": "使用用户名密码登录",
                "operationId": "login",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "用户名",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "密码",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/definitions/User"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/current-user": {
            "get": {
                "tags": [
                    "Common"
                ],
                "summary": "获取当前用户解析",
                "description": "获取当前用户解析",
                "operationId": "current-user",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    
                ],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/definitions/User"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "新建用户",
                "description": "添加一个用户",
                "operationId": "createUser",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "创建一个用户对象",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            },
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "获取用户列表",
                "description": "获取用户列表",
                "operationId": "getUserList",
                "produces": [
                    "application/json"
                ],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/user/{id}": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "删除单个用户",
                "description": "根据用户id删除用户",
                "operationId": "removeUserById",
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "用户id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/definitions/User"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "summary": "修改单个用户",
                "description": "根据用户id修改用户",
                "operationId": "updateUserById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "用户id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "创建一个用户对象",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/definitions/User"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "User"
                ],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "summary": "获取单个用户",
                "description": "根据用户id获取用户",
                "operationId": "getUserById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "用户id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/definitions/User"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "api_key": {
            "type": "apiKey",
            "name": "authorization",
            "in": "header"
        }
    },
    "definitions": {
        "User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64",
                    "description": "用户id"
                },
                "wx_name": {
                    "type": "string",
                    "description": "微信名"
                },
                "avatar": {
                    "type": "string",
                    "description": "头像地址"
                },
                "name": {
                    "type": "string",
                    "description": "自定义名称"
                },
                "role": {
                    "type": "string",
                    "description": "角色"
                },
                "init": {
                    "type": "string",
                    "description": "是否初始化"
                },
                "slot": {
                    "type": "string",
                    "description": "随机盐"
                },
                "password": {
                    "type": "string",
                    "description": "密码"
                }
            }
        }
    },
    "externalDocs": {
        "description": "Find out more about Swagger",
        "url": "http://swagger.io"
    }
}